Refactor the BasePredictor
WIP
